Map_metabolites_iKS1317_Sco4
Excell table: Mapping of metabolites from Sco4 to iKS1317 in TABLE1 (red are involved in both models, brown are manually checked and they present the same metabolites in both models-although they have different IDs in the two models, blue are in iKS1317 but not in Sco4 and green bold are in Sco4 but not in iKS1317). So the green metabolites should be added to iKS1317-those metabolites I mapped also on KEGG ID. Each metabolites have A-S columns.