Mapping Reactions iKS1317 Sco4
Here is the mapping of reactions from Sco4 to iKS1317 in TABLE1 (red are involved in both models, brown are manually checked and they present the same reactions in both models-although they have different IDs in the two models, blue are in iKS1317 but not in Sco4 and green bold are in Sco4 but not in iKS1317). So the green reactions should be added to iKS1317. Each reaction have A-AA columns.