Adds unfinished scripts and csv-files
The two csv files S4 and S5 corresponds to the supplementary data in the
paper related to iAA1259. S4 contains the list of new reactions added in
iAA1259. This list is now curated and I found 4-5 reactoins that were
not new after all. This csv file (S4) will be used to add reactions from
iAA1259 to scoGEM.

Note that they have used BIGG ids in sAA1259, and we can use the
overlap to map some reactions added from Sco4 to BIGG ids