feat: wild-type growth tests
Adds a test for wild-type growth tests.
Currently this test PASS if the growth predictions
are similar or better than the results from iKS1317.
The data table wild_type.csv was also updates, because there were some
errors in this file.

The current results are similar as the iKS1317-results:

- TP: 51
- FN: 0
- FP: 2 (aspartate and glutamine as carbon sources)
- TN: 4

The cut-off value for the growth rate is 0.028, corresponding to a
doubling time of 24 hours. The uptake rate of carbon is scaled for each
carbon source to correspond to the maximum observed growth rate for
glucose: 2.1 mml/gDW/h, i.e. 6*21 mmol carbon atoms/gDW/h. Nitrogen
uptake is scaled similarly to max 1.85 mmol/gDW/h.