feat: growth test for mutants from litterature
Adds a test which compare in silico and in vivo growth
for knock-out mutants found in the litterature.
Currently this is just on test which PASS if
the predicitions are better (more correct) than
the predictions from iKS1317.

The test is currently failing because of the following discrepancies:
- Growth on maltose for the delta malE mutant
- Growth on L-valine for the delta vdh mutant
- Growth with nitrate as nitrogen source for the delta nirB mutant

Additionally, the model predicts erronously no-growth for the delta
msdA mutant on propionate, however this is identical to iKS1317,
and caused by the coupling introduced in iKS1317.