feat: growth test for mutants from transposon study
Created a test which compare in silico and an in vivo
growth for knock-out mutants from the transposon
mutagenesis study by xu et al.
The test PASS if the number pf predictions is better than iKS1317
in all categories (TP, FP, FN, TN), however this should maybe be
modified.

Currently the test fails because of an increased number of false
positives.

Additional change: Added a metric to all growth tests (accuracy)