refactor: commenting out unused code
The
`find_internal_and_unrealistic_energy-generating_cycles_scoGEM.py`
didn't work for me so I took the freedom to clean up the script a bit
and commented out unused code and import of unused packages.

I didn't remove any code, because I wasn' sure if it was supposed to be
used for something else. A quick summary of the changes I made:
- Uncommented unused code
- The script used find_egc which is a static function. I copied this out
of the class stFBA so it could be used without creating an instance of
that class (which wasn't used for anythng else).
- Small changes in how to set the solver made the code run with cobra
0.13.4
- Uncommente the import of a settings file and hardcoded in the solver.
Currently it doesn't work with gurobi, but cplex runs fine

I think it would be nice to include the find_energy_generating_cycles in
the tests, I'm creating an issue on it.