docs: python requirements
Adds a function to the export.py scripts which write the text file
requirements.txt which describe the additional packages (and
versions) outside standard python packages required to run the model
reconstruction. The requirements file is based on the actual imports
in all scripts, and will therefore be updated automatically if we
use any new packages in any of the scripts. These packages can be
installed using 'pip install -r requirements.txt'

This solves the python part of #46

Also updates the README.md file