feat: make models with changed reversibility
The script reversibility.py create models where the reversibility
of some reactions are changed according to the data files / input.

Summary of results for equilibrator model:
reversible to forward: 24
forward to reversible: 300
reversible to backward: 16
forward to backward: 20
Changed the reversibility of 360 reactions in total. Data were available
for 696 reactions

Summary of results from metacyc model:
reversible to forward: 97
forward to reversible: 91
Changed the reversibility of 188 reactions in total. Data were available
for 740 reactions