fix/refactor: removed model import
When running memote the model is already loaded into the session,
and by removing the explicit import in these two scripts we use the
model selected when running memote.

Not that the tests won't work running pytest.

The range for acceptable growth rate was increased to 7.1 as the lower
bound (7.3 previously)