feat/fix: updated rates for germicidin and red
After a brief talk with Alexander we decided to use the linear phase of
the undecylprodigiosin production to estimate the rate instead of
fitting a sigmoid. The reason is that the appareant saturation from the
MS-data does not correspond to photometric (?) measurements which show
no saturation.

Additionally, the germicidin B rate for M1152 were updated (bug fix)