feat: increased coverage of eQuilibrator dGs
Adds a csv-file with delta Gm' and delta G0' values extracted
automatically from eQuilibrator using the python API by Elad Noor.
The increases the coverage from 697 to 887. All dG values are according
to the direction of reaction in the model. The data is calculated using
pH = 7 and ionic strength = 0.1

The script use less than a minute, so it can easily be rerun with
different parameters