feat: eQuilibrator model updated
After grabbing new dG values from eQuilibrator autmatically with the
python API the coverage increase from 695 isj to 876 reactions.
A screen of "lethal" interaction of changes are stored in
eQulibrator_reversibility_lethals.csv.

With "lethal" we mean effects that change the biomass growth more than a
pre-defined threshold in either directions: +- 0.01 /h was used.
The reaction directionality was determined using +- 30 as the threshold
for dGm_prime, withour taking the standard deviation intpo account.

Of the 876 reactions with dG estimated, 392 suggested change in
reversibility. All of these were inccorporated in  eqscoGEM.xml, except
any of the ractions in one or more of the "lethal" pairs given in
eQulibrator_reversibility_lethals.csv. In total 321 reactions were
changed.