docs: adds misc files for completeness
There are a couple of files / scripts that were generated / used to
generate mappings etc during the reconstruction process, but is actually
not a part of the pipeline for reconstruction. This is typically mapping
of models to create a csv-file, which is then manually curated and used
in the reconstruction pipeline. This also includes the iMK1208 model.
Some paths needed to be updated in one script to reflect these files
added to the repository.