feat/refactor: RED/germ update pluss format
feat: had to discard the first data point from estimating the rate of
germicidin for M1152and the last data points from RED for M145 for fitting the rates.

refactor: Limited the number of decimal places to four. Also changed
from semicolon to comma as the seperator to make the file readable in
gitHub