integration_transp_rev_scoGEM
In the scoGEM, I have integrated the transport rxns, as well as exchange rxns and sink rxns with respect tp transport rxns. Transport rxns were devided into 3 tables with prefix Table1, Table1, Table3, inlcuding existing transport rxns with updated/changed gene association (prefix Table1), new transport rxns (prefix Table2), and new transport rxns with new metabolites(prefix Table3), Rxns that include exchnage and sink rxns are in uploaded in tables with prefix Table4 and Table5, In addition to adding the transporters, rxns with rxnIDs of VALTA, LEUTA and ILETA were changed from reversible to ireversible in order to satisfy in vivo growth data on knockout (SCO4089) with L-leucine, L-valine, L-isoleucine  as N-source and D-glucose as C source. This final scoGEM model has ben evaluated  with respect to wild type growth on all carbon sources and  all nitrogen sources in the model and for knockout growth and compared to qualitative in vivo data, growth data on the model iKS1317 and growth data on iMK1208.