scoGEM_added_transporters
To scoGEM model, transport rxns were addedm with existing metabolites already in scoGEM model and with new metbaolites, for which sink and exchange rxns were added. In the scoGEM model for each of the transpor rxns, RxnID, RxnFormula, Gene Association, and lower and upper bound were integrated. Corresponding tables describing the integrated transporters are uploaded and integrated in the uploaded script.

1. Table1_updated_grRules_280119_300119_6.csv (updated gene association on existing transport rxns)
2. Table2_newTransportRxns_280119_300119_3.csv (new transporters with existing metabolites)
3. Table3_newTransportRxns_newMetabolites_280119_300119_3.csv (new transporters with new transporters)
4. ExRxns_added_Transporters_300119_3.csv (added exchange rxns with respect to new metabolites)
5. Sink_Rxns_scoGEM_Transporters_300119_3.csv (added sink rxns with respect to new metabolites)

In addition reversibility  rxns with RxnID: VALTA, LEUTA, ILETA was changed to irreversible.