fix/refactor: mmy metabolite name issue
Because of some issue with the propagation of the metabolite ids with
optlang and gurobipy (https://github.com/opencobra/cobrapy/issues/819)
the reconstruct_scoGEM.py doesn't work with gurobi as the solver.
Therefore, glpk is used as solver in the reconstruction of ScoGEM, these
changes are incorporated into reconstruct_scoGEM.py

Additionally, this revealed an issue that was adressed but not solved properly in
fix_ikS1317_issues.py, but solved in fix_issue33_annotation_bugs.py.
This was very confusing. Thus, this issue is now sorted out completely
in fix_iKS1317_issues.py, and removed from
fix_issue33_annotation_bugs.py.

Finally, I've upgraded cobrapy to 0.14.2 in the requirements, as this
was wrongly suggested to cause the issued described above (see also PR
 #73)