Fix/transporters (#97)
Applies #95 to all new reactions (some were accidentially left out in #95).

* fix: exchange-reactions lower bound to 0

The exchange reactions added to new metabolites were automatically
given a lower bound of -1000. This is now fixed so that the lower
bound is 0.

Additionally, SBO-annotations was changed to lower case.

* fix: exchange-reactions lower bound to 0

The exchange reactions added to new metabolites were automatically
given a lower bound of -1000. This is now fixed so that the lower
bound is 0.